@@ -18,7 +18,7 @@ dependencies:
   eva_icons_flutter: ^2.0.1
   flutter_spinkit: ^4.1.2+1
   shared_preferences: ^0.5.12+2
-  google_maps_flutter: ^1.0.6
+  google_maps_flutter: ^1.2.0
   http: ^0.12.2
   permission_handler: ^5.0.1+1
   bottom_sheet: ^0.0.1-dev.7
@@ -37,7 +37,7 @@ dependencies:
   flutter_svg_provider: ^0.1.8
   flutter_screenutil: ^4.0.3+2
   mask_text_input_formatter: ^1.2.1
-  get: ^3.25.6
+  qr_code_scanner:
 
 dev_dependencies:
   flutter_test:
@@ -59,6 +59,7 @@ flutter:
     - assets/icons/onboarding_4/People.svg
     - assets/icons/onboarding_4/Clouds.svg
     - assets/svg/
+    - svg/
 
   uses-material-design: true
   fonts:
@@ -74,3 +75,6 @@ flutter:
     - family: LofOut
       fonts:
         - asset: assets/fonts/LofOut.ttf
+    - family: TabIcons
+      fonts:
+        - asset: assets/fonts/TabIcons.ttf
