@@ -13,9 +13,13 @@ InputDecoration inputDecorWidget() {
           borderRadius: BorderRadius.all(Radius.circular(25)),
           borderSide: BorderSide(color: Color(0xFF62C848), width: 2.5)),
       disabledBorder: OutlineInputBorder(
-          borderSide: BorderSide(color: Color(0xFF62C848), width: 2.5)),
+        borderSide: BorderSide(color: Color(0xFF62C848), width: 2.5),
+        borderRadius: BorderRadius.all(Radius.circular(25)),
+      ),
       focusedBorder: OutlineInputBorder(
+          borderRadius: BorderRadius.all(Radius.circular(25)),
           borderSide: BorderSide(color: Color(0xFF62C848), width: 2.5)),
       border: OutlineInputBorder(
+          borderRadius: BorderRadius.all(Radius.circular(25)),
           borderSide: BorderSide(color: Color(0xFF62C848), width: 2.5)));
 }
