@@ -13,9 +13,6 @@ import 'package:recycle_hub/screens/tabs/map/widgets/loader_widget.dart';
 import 'package:recycle_hub/style/theme.dart';
 import 'package:flutter_typeahead/flutter_typeahead.dart';
 
-GarbageCollectionTypeBloc garbageCollBloc = GarbageCollectionTypeBloc();
-MarkerWorkModeBloc markerWorkModeBloc = MarkerWorkModeBloc();
-
 class MapFilterDetailScreen extends StatefulWidget {
   @override
   _MapFilterDetailScreenState createState() => _MapFilterDetailScreenState();
@@ -27,6 +24,8 @@ class _MapFilterDetailScreenState extends State<MapFilterDetailScreen> {
   FilterModel currentFilterModel = FilterModel();
   AcceptTypesCollection acceptTypesCollection;
   List<FilterCardWidget> filterCards;
+  GarbageCollectionTypeBloc garbageCollBloc = GarbageCollectionTypeBloc();
+  MarkerWorkModeBloc markerWorkModeBloc = MarkerWorkModeBloc();
   /*GlobalKey<FilterCardWidgetState> _key =
       GlobalKey<FilterCardWidgetState>(debugLabel: "__myKey__");*/
 
@@ -47,10 +46,17 @@ class _MapFilterDetailScreenState extends State<MapFilterDetailScreen> {
         }
       },
       child: Scaffold(
-          backgroundColor: Color(0xFFFFFFFF),
+          backgroundColor: kColorWhite,
           appBar: AppBar(
             title: Text("Фильтр"),
             centerTitle: true,
+            leading: GestureDetector(
+              onTap: () {
+                markersCollectionBloc.loadMarkers();
+                Navigator.pop(context);
+              },
+              child: Icon(Icons.arrow_back),
+            ),
           ),
           body: Container(
             color: Color(0xFFFFFFFF),
@@ -59,25 +65,31 @@ class _MapFilterDetailScreenState extends State<MapFilterDetailScreen> {
               children: [
                 Padding(
                   padding: const EdgeInsets.fromLTRB(15, 5, 15, 5),
-                  child: TypeAheadField(
-                    textFieldConfiguration: TextFieldConfiguration(
-                      controller: _searchController,
-                      decoration: inputDecorWidget(),
+                  child: Container(
+                    decoration: BoxDecoration(
+                        color: Color(0xFFF2F2F2),
+                        borderRadius: BorderRadius.all(Radius.circular(10))),
+                    child: TypeAheadField(
+                      textFieldConfiguration: TextFieldConfiguration(
+                        controller: _searchController,
+                        decoration: inputDecorWidget(),
+                        cursorColor: kColorBlack,
+                      ),
+                      suggestionsCallback: (str) {
+                        return acceptTypesCollection.getPatterns(str);
+                      },
+                      itemBuilder: (BuildContext context, suggestion) {
+                        return ListTile(
+                          title: Text(suggestion),
+                        );
+                      },
+                      onSuggestionSelected: (String suggestion) {
+                        selectCardByVarName(acceptTypesCollection
+                            .getVarNameByKeyWord(suggestion));
+                        currentFilterModel.filters.add(acceptTypesCollection
+                            .getVarNameByKeyWord(suggestion));
+                      },
                     ),
-                    suggestionsCallback: (str) {
-                      return acceptTypesCollection.getPatterns(str);
-                    },
-                    itemBuilder: (BuildContext context, suggestion) {
-                      return ListTile(
-                        title: Text(suggestion),
-                      );
-                    },
-                    onSuggestionSelected: (String suggestion) {
-                      selectCardByVarName(acceptTypesCollection
-                          .getVarNameByKeyWord(suggestion));
-                      currentFilterModel.filters.add(acceptTypesCollection
-                          .getVarNameByKeyWord(suggestion));
-                    },
                   ),
                 ),
                 Padding(
@@ -225,12 +237,12 @@ class _MapFilterDetailScreenState extends State<MapFilterDetailScreen> {
   _markerModeCheck() {
     return StreamBuilder<Object>(
         stream: markerWorkModeBloc.stream,
-        initialData: markerWorkModeBloc.defaultItem,
+        //initialData: markerWorkModeBloc.defaultItem,
         builder: (context, snapshot) {
           return Container(
               height: 60,
               decoration: BoxDecoration(
-                  borderRadius: BorderRadius.circular(5),
+                  borderRadius: BorderRadius.circular(2),
                   border: Border.all(color: Color(0xFF62C848), width: 1.5)),
               child: Row(
                 mainAxisAlignment: MainAxisAlignment.center,
@@ -340,7 +352,7 @@ class _MapFilterDetailScreenState extends State<MapFilterDetailScreen> {
   _garbCollectTypeWidget() {
     return StreamBuilder(
       stream: garbageCollBloc.stream,
-      initialData: garbageCollBloc.defaultItem,
+      //initialData: garbageCollBloc.defaultItem,
       builder: (ctx, AsyncSnapshot<GCOLLTYPE> snapshot) {
         return Container(
           height: 40,
@@ -433,21 +445,26 @@ class _MapFilterDetailScreenState extends State<MapFilterDetailScreen> {
 
   InputDecoration inputDecorWidget() {
     return InputDecoration(
-        prefixIcon: Icon(
-          Icons.search,
-          color: Color(0xFF62C848),
-          size: 40,
+        focusColor: Color(0xFFF2F2F2),
+        fillColor: Color(0xFFF2F2F2),
+        hoverColor: Color(0xFFF2F2F2),
+        suffixIcon: Padding(
+          padding: const EdgeInsets.fromLTRB(10, 5, 15, 5),
+          child: Icon(
+            Icons.search,
+            color: Color(0xFF616161),
+            size: 35,
+          ),
         ),
         hintText: "Что вы хотите сдать?",
-        hintStyle: TextStyle(color: Color(0xFF62C848), fontSize: 16.0),
+        hintStyle: TextStyle(
+          color: Color(0xFF616161).withOpacity(0.6),
+          fontSize: 16.0,
+        ),
+        prefix: SizedBox(width: 15),
         enabledBorder: OutlineInputBorder(
-            borderRadius: BorderRadius.all(Radius.circular(25)),
-            borderSide: BorderSide(color: Color(0xFF62C848), width: 2.5)),
-        disabledBorder: OutlineInputBorder(
-            borderSide: BorderSide(color: Color(0xFF62C848), width: 2.5)),
-        focusedBorder: OutlineInputBorder(
-            borderSide: BorderSide(color: Color(0xFF62C848), width: 2.5)),
-        border: OutlineInputBorder(
-            borderSide: BorderSide(color: Color(0xFF62C848), width: 2.5)));
+            borderRadius: BorderRadius.all(Radius.circular(10)),
+            borderSide: BorderSide.none),
+        focusedBorder: OutlineInputBorder(borderSide: BorderSide.none));
   }
 }
